requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.3
pdf2image==1.17.0
Pillow==10.3.0
pytesseract==0.3.10
python-dotenv==1.0.1
watchdog==4.0.0
markdown==3.9
